requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
Django==4.2.7
idna==3.4
imaginesdk @ git+https://github.com/Vyro-ai/imagine-sdk-python.git@4cf43c3fdde5286eb30885ffa0ca3aafcf57c3aa
Pillow==10.1.0
python-dotenv==1.0.0
requests==2.31.0
sqlparse==0.4.4
typing-extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
